% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmetrik.R
\docType{package}
\name{jmetrik}
\alias{jmetrik}
\alias{jmetrik-package}
\title{Tools for interacting with 'jMetrik'}
\description{
\code{jmetrik} provides tools for using R and the external program 'jMetrik' togther. In particular, it
provides tools for reading and writing files in *.jmetrik format. This format is required by 'jMetrik'
and it is a flat file with a header and comma separated values. However, a *.jmetrik file is not simply
a CSV file. The header includes much more information about the data than is typically found in a CSV file.
'jMetrik' is an open source Java application for psychometric analysis. It may be downloaded from
\url{http://www.ItemAnalysis.com}.
}
\details{
See \code{\link{jmetrikWrite}} and \url{http://www.ItemAnalysis.com} for more information about the *.jmetrik
file format.
}
