% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jmetrik.R
\name{jmetrikRead}
\alias{jmetrikRead}
\title{Reads a *.jmetrik file into a data frame.}
\usage{
jmetrikRead(fileName)
}
\arguments{
\item{fileName}{The complete path and file name of the *.jmetrik file that is being read.}
}
\value{
a data frame
}
\description{
A *.jmetrik file can be created with \code{\link{jmetrikWrite}} or by the 'jMetrik' program.
See \url{http://www.ItemAnalysis.com}.
}
\examples{
\dontrun{
x<-jmetrikRead(fileName="./myfile.jmetrik")
}
}

