% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmetrik.R
\name{jmetrikRead}
\alias{jmetrikRead}
\title{Reads a *.jmetrik file into a data frame.}
\usage{
jmetrikRead(fileName, maxScan = 500)
}
\arguments{
\item{fileName}{The complete path and file name of the *.jmetrik file that is being read.}

\item{maxScan}{The maximum number of rows to scan. This number should be at least the number
of variables int eh data file.}
}
\value{
a data frame
}
\description{
A *.jmetrik file can be created with \code{\link{jmetrikWrite}} or by the 'jMetrik' program. 
See \url{http://www.ItemAnalysis.com}.
}
\examples{
\donttest{
x<-jmetrikRead(fileName=system.file("extdata", "exam1iparam.jmetrik", package = "jmetrik"))
}

}
